// src/interfaces/Movies.ts
export interface MovieResponse {
    movies: Movie[];
}

export interface MovieDetail {
    movie: Movie;
}
export interface Movie {
    _id:           string;
    title:         string;
    director:      string;
    overview:      string;
    backdrop_path?: string;
    poster_path?:   string;
    actors:        string;
    release_date:  string;
    trailer: string
}
// Generated by https://quicktype.io